release 3.1.38 source code for java

+++ services/rocketmq/src/main/java/com/huaweicloud/sdk/rocketmq/v2/model/ConsumerList.java
@@ -3,7 +3,6 @@ package com.huaweicloud.sdk.rocketmq.v2.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -22,7 +21,7 @@ public class ConsumerList {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "total")
 
-    private BigDecimal total;
+    private Integer total;
 
     public ConsumerList withTopics(List<String> topics) {
         this.topics = topics;
@@ -57,7 +56,7 @@ public class ConsumerList {
         this.topics = topics;
     }
 
-    public ConsumerList withTotal(BigDecimal total) {
+    public ConsumerList withTotal(Integer total) {
         this.total = total;
         return this;
     }
@@ -66,11 +65,11 @@ public class ConsumerList {
      * Topic总数（当查询topic消费“列表”时才显示此参数）。
      * @return total
      */
-    public BigDecimal getTotal() {
+    public Integer getTotal() {
         return total;
     }
 
-    public void setTotal(BigDecimal total) {
+    public void setTotal(Integer total) {
         this.total = total;
     }
