+++ services/rocketmq/src/main/java/com/huaweicloud/sdk/rocketmq/v2/model/ShowConsumerListOrDetailsResponse.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.huaweicloud.sdk.core.SdkResponse;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -23,7 +22,7 @@ public class ShowConsumerListOrDetailsResponse extends SdkResponse {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "total")
 
-    private BigDecimal total;
+    private Integer total;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "brokers")
@@ -63,7 +62,7 @@ public class ShowConsumerListOrDetailsResponse extends SdkResponse {
         this.topics = topics;
     }
 
-    public ShowConsumerListOrDetailsResponse withTotal(BigDecimal total) {
+    public ShowConsumerListOrDetailsResponse withTotal(Integer total) {
         this.total = total;
         return this;
     }
@@ -72,11 +71,11 @@ public class ShowConsumerListOrDetailsResponse extends SdkResponse {
      * Topic总数（当查询topic消费“列表”时才显示此参数）。
      * @return total
      */
-    public BigDecimal getTotal() {
+    public Integer getTotal() {
         return total;
     }
 
-    public void setTotal(BigDecimal total) {
+    public void setTotal(Integer total) {
         this.total = total;
     }
